requirements.txt
pyhocon
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
pyyaml==5.2
scikit-learn==0.19.1
scipy
torch==1.2.0
